+++ generators/app/templates/es/.eslintrc.js
@@ -57,9 +57,7 @@ module.exports = {
     'import/dynamic-import-chunkname': ERROR,
     'import/extensions': [ERROR, 'always', {
       js: 'never',
-      jsx: 'never',
-      ts: 'never',
-      tsx: 'never'
+      jsx: 'never'
     }],
     'import/no-extraneous-dependencies': OFF,
     'import/no-dynamic-require': OFF,
@@ -97,47 +95,7 @@ module.exports = {
     'react/no-unused-state': ERROR,
     'react/prefer-stateless-function': ERROR,
     'react/require-default-props': OFF,
-    'react/sort-comp': [ERROR, {
-      order: [
-        'props',
-        'static-methods',
-        'lifecycle',
-        'everything-else',
-        'handlers',
-        'rendering',
-      ],
-      groups: {
-        props: [
-          'propTypes',
-          'defaultProps',
-        ],
-        lifecycle: [
-          'statics',
-          'constructor',
-          'getDerivedStateFromProps',
-          'componentWillMount',
-          'UNSAFE_componentWillMount',
-          'componentDidMount',
-          'componentWillReceiveProps',
-          'UNSAFE_componentWillReceiveProps',
-          'shouldComponentUpdate',
-          'componentWillUpdate',
-          'UNSAFE_componentWillUpdate',
-          'getSnapshotBeforeUpdate',
-          'componentDidUpdate',
-          'componentDidCatch',
-          'componentWillUnmount',
-        ],
-        handlers: [
-          '/^on.+$/',
-          '/^handle.+$/',
-        ],
-        rendering: [
-          '/^render.+$/',
-          'render',
-        ],
-      }
-    }],
+    'react/sort-comp': ERROR,
     'react/static-property-placement': OFF,
     'require-await': ERROR
   },