+++ generators/app/templates/es/.postcssrc.js
@@ -1,5 +1,7 @@
+const postcssPresetEnv = require('postcss-preset-env');
+
 module.exports = {
-  plugins: {
-    'autoprefixer': {}
-  }
+  plugins: [
+    postcssPresetEnv()
+  ]
 }